+++ CHANGELOG.md
@@ -41,3 +41,4 @@ a breaking changes version
 - Add testing for File.watch() and Dir.watch() and unwatch
 - Adding private property "watcher" to File
 - Changing API for File.watch() and Dir.watch()
+- Fixing Bug found in Shape.createShapeInst()